# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API untuk sistem HRIS (Human Resource Information System)",
        "title": "HRIS API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/divisi": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Ambil daftar semua divisi",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Divisi"
                ],
                "summary": "Get semua divisi",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Tambah divisi baru",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Divisi"
                ],
                "summary": "Buat divisi",
                "parameters": [
                    {
                        "description": "Divisi Request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.DivisiRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/divisi/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update nama divisi berdasarkan ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Divisi"
                ],
                "summary": "Update divisi",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Divisi ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Divisi Request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.DivisiRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Hapus divisi berdasarkan ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Divisi"
                ],
                "summary": "Delete divisi",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Divisi ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/employees": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Ambil semua data detail karyawan",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Employees"
                ],
                "summary": "Get semua data karyawan",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/employees/{uid}/data": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Ambil data detail karyawan berdasarkan UID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Employees"
                ],
                "summary": "Get data user by UID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "uid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Simpan data detail karyawan",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Employees"
                ],
                "summary": "Buat data user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "uid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Data User Request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.DataUserRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.DataUserRequest": {
            "type": "object",
            "properties": {
                "alamat_ktp": {
                    "type": "string"
                },
                "alamat_tinggal": {
                    "type": "string"
                },
                "divisi_id": {
                    "type": "string"
                },
                "domisili_ktp": {
                    "type": "string"
                },
                "domisili_lahir": {
                    "type": "string"
                },
                "domisili_tinggal": {
                    "type": "string"
                },
                "is_bpjs_tk": {
                    "type": "boolean"
                },
                "jabatan_id": {
                    "type": "string"
                },
                "jenis_asuransi": {
                    "type": "string"
                },
                "nama_asuransi": {
                    "type": "string"
                },
                "no_asuransi": {
                    "type": "string"
                },
                "no_bpjs": {
                    "type": "string"
                },
                "no_npwp": {
                    "type": "string"
                },
                "no_rekening": {
                    "type": "string"
                },
                "nomor_telepon": {
                    "type": "string"
                },
                "nomor_telepon_darurat": {
                    "type": "string"
                },
                "nomor_telepon_second": {
                    "type": "string"
                },
                "prov_ktp": {
                    "type": "string"
                },
                "prov_tinggal": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "status_perkawinan": {
                    "type": "string"
                },
                "tanggal_lahir": {
                    "type": "string"
                },
                "tanggal_masuk": {
                    "type": "string"
                },
                "tipe_bank": {
                    "type": "string"
                },
                "tipe_tinggal": {
                    "type": "string"
                }
            }
        },
        "model.DivisiRequest": {
            "type": "object",
            "required": [
                "name_divisi"
            ],
            "properties": {
                "name_divisi": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}